requirements.txt
PyYAML==3.13
requests==2.19.1
PyMySQL==0.9.2
djangorestframework==3.8.2
redis==2.10.6
Django==2.1
pyecharts==0.5.8
echarts-themes-pypkg==0.0.3
django-redis==4.9.0
django-redis-cache==1.7.1
